requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.4.0
drf-yasg==1.21.9
pytest==8.3.4
python-dotenv==1.0.1
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0
django-cors-headers==4.3.1
django-jazzmin==2.4.0
requests==2.31.0
aiogram==3.4.1
sqlalchemy==2.0.29
whitenoise==6.6.0
Pillow==9.0.0
